Add post function 投稿ページを進めました。 Pushed by AKI
